% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_search.R
\name{.embedding_rerank_enabled}
\alias{.embedding_rerank_enabled}
\title{Check if embedding rerank is enabled}
\usage{
.embedding_rerank_enabled()
}
\description{
Check if embedding rerank is enabled
}
\keyword{internal}
