% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary-helpers.R
\name{infer_column_role}
\alias{infer_column_role}
\title{Infer column role from name and data}
\usage{
infer_column_role(col_name, col)
}
\arguments{
\item{col_name}{Column name}

\item{col}{Column vector}
}
\value{
Character string indicating the column role
}
\description{
Infer column role from name and data
}
\keyword{internal}
