% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary-helpers.R
\name{infer_value_type}
\alias{infer_value_type}
\title{Infer value type from a column}
\usage{
infer_value_type(col)
}
\arguments{
\item{col}{A column vector}
}
\value{
Character string indicating the value type
}
\description{
Infer value type from a column
}
\keyword{internal}
