% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_validate.R
\name{print.ms_validation_report}
\alias{print.ms_validation_report}
\title{Print method for ms_validation_report}
\usage{
\method{print}{ms_validation_report}(x, ...)
}
\arguments{
\item{x}{An \code{ms_validation_report} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for ms_validation_report
}
