% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-deduplication.R
\name{suggest_facet_schemes}
\alias{suggest_facet_schemes}
\title{Suggest facet schemes for proposed terms}
\usage{
suggest_facet_schemes(proposed_terms)
}
\arguments{
\item{proposed_terms}{A data frame with term_label column}
}
\value{
A tibble with suggested facet schemes and their member concepts
}
\description{
Analyzes a proposed terms dataframe and suggests which facet schemes
(age class, life phase, etc.) should be created instead of proliferating
individual terms.
}
\examples{
proposed <- tibble::tibble(
  term_label = c("Spawner Age 1", "Spawner Age 2", "Spawner Age 3")
)
facets <- suggest_facet_schemes(proposed)
facets$scheme_name
}
