% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metasalmon-package.R
\docType{package}
\name{metasalmon}
\alias{metasalmon-package}
\alias{metasalmon}
\title{metasalmon: Salmon Monitoring Utilities}
\description{
Package scaffold for ingesting, validating, transforming, summarizing, and
visualizing salmon monitoring datasets. Functions and data will be added
incrementally.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dfo-pacific-science.github.io/metasalmon/}
}

}
\author{
\strong{Maintainer}: Brett Johnson \email{brettthomasjohnson@gmail.com}

}
\keyword{internal}
