% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metasalmon-package.R
\docType{package}
\name{metasalmon}
\alias{metasalmon-package}
\alias{metasalmon}
\title{metasalmon: Utilities for Salmon Data Packages}
\description{
Tools to scaffold, standardize, validate, transform, and package salmon
datasets using the Salmon Data Package conventions plus shared and
DFO-specific salmon vocabularies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://salmon-data-mobilization.github.io/metasalmon/}
  \item Report bugs at \url{https://github.com/salmon-data-mobilization/metasalmon/issues}
}

}
\author{
\strong{Maintainer}: Brett Johnson \email{brettthomasjohnson@gmail.com}

}
